Refactor Ticket Service to Apply SOLID Principles
- Moved logic for Freshdesk and database ticket retrieval into the TicketService.
- Implemented validation, ticket fetching, and CRUD operations while aligning with SOLID principles.
- Refactored controller to focus on handling HTTP requests and responses only.
- Enhanced  method to support both Freshdesk ID and database ID.
- Ensured proper separation of concerns by using Service Layer for business logic.
